requirements.txt
chromadb>=0.5.0
pydantic>=2.0.0
google-generativeai>=0.8.0
streamlit>=1.40.0
supabase>=2.0.0
python-dotenv>=1.0.0
